cto_signal_scanner/requirements.txt
flask==3.0.0
flask-wtf==1.2.1
flask-talisman==1.1.0
python-dotenv==1.0.0
openai==1.3.0
feedparser==6.0.10
reportlab==4.0.7
requests==2.31.0
python-dateutil==2.8.2 